Refactor Weeder
